+++ CHANGELOG.md
@@ -12,6 +12,9 @@ about writing changes to this log.
 
 - [PR-366](https://github.com/itk-dev/naevnssekretariatet/pull/366)
   Removed `InspectionLetter` and its logic.
+- [PR-364](https://github.com/itk-dev/naevnssekretariatet/pull/364)
+  Removes button for creating `HearingPostResponse` and `Decision`,
+  and receipt logic.
 
 ## [1.4.1] 2023-04-25