+++ CHANGELOG.md
@@ -12,6 +12,22 @@ about writing changes to this log.
 
 - [PR-356](https://github.com/itk-dev/naevnssekretariatet/pull/356)
   Updated hearing data structure to allow multiple recipients
+- [PR-367](https://github.com/itk-dev/naevnssekretariatet/pull/367)
+  Preselected party relation type upon edit
+- [PR-363](https://github.com/itk-dev/naevnssekretariatet/pull/363)
+  Validate party identifier based on identifier type
+- [PR-366](https://github.com/itk-dev/naevnssekretariatet/pull/366)
+  Removed `InspectionLetter` and its logic.
+- [PR-364](https://github.com/itk-dev/naevnssekretariatet/pull/364)
+  Removes button for creating `HearingPostResponse` and `Decision`,
+  and receipt logic.
+
+## [1.4.1] 2023-04-25
+
+- [PR-360](https://github.com/itk-dev/naevnssekretariatet/pull/360)
+  Fixed type.
+- [PR-361](https://github.com/itk-dev/naevnssekretariatet/pull/361)
+  Update accordions
 
 ## [1.4.0] 2023-04-25
 
@@ -253,7 +269,8 @@ Fixed error in unescaped characters in filename
 - [TVIST1-604](https://jira.itkdev.dk/browse/TVIST1-604):
   Resolved issue regarding time formats.
 
-[Unreleased]: https://github.com/itk-dev/naevnssekretariatet/compare/1.4.0...HEAD
+[Unreleased]: https://github.com/itk-dev/naevnssekretariatet/compare/1.4.1...HEAD
+[1.4.1]: https://github.com/itk-dev/naevnssekretariatet/compare/1.4.0...1.4.1
 [1.4.0]: https://github.com/itk-dev/naevnssekretariatet/compare/1.3.2...1.4.0
 [1.3.2]: https://github.com/itk-dev/naevnssekretariatet/compare/1.3.1...1.3.2
 [1.3.1]: https://github.com/itk-dev/naevnssekretariatet/compare/1.3.0...1.3.1